shadowing mode (-s) never produced shadows... clSetKernelArg is used to pass zmax to the kernel. while zmax was declared as double zmax inside the kernel was always 0.0! declareing it as float fixes this BUT shadowing produces strange artefacts :(

+++ raster/r.sun/rsun_opencl.h
@@ -72,7 +72,7 @@ struct OCLConstants {
 	int refl_rad;
 	int glob_rad;
     double degreeInMeters;
-    double zmax;
+    float zmax;
     
     float linke_max, linke_min;
     float albedo_max, albedo_min;
